[BUGFIX] Prevent PHP warning when saving in BE

+++ Classes/Service/AssetService.php
@@ -656,6 +656,9 @@ class Tx_Vhs_Service_AssetService implements t3lib_Singleton {
 			return;
 		}
 		$assetCacheFiles = glob(t3lib_div::getFileAbsFileName('typo3temp/vhs-assets-*'));
+		if (FALSE === $assetCacheFiles) {
+			return;
+		}
 		foreach ($assetCacheFiles as $assetCacheFile) {
 			unlink($assetCacheFile);
 		}
